move inf scroll and virtualization into own component
